Combined the Rpc classes into the base Message queue system and removed associated projects.
