--- DtronixMessageQueue.Rpc.Tests/SimpleRpcSession.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DtronixMessageQueue.Rpc.Tests {
-	public class SimpleRpcSession : RpcSession<SimpleRpcSession> {
-	}
-}
