--- DtronixMessageQueue.Rpc.Tests/packages.config
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net45" />
-  <package id="xunit" version="2.2.0-beta2-build3300" targetFramework="net45" />
-  <package id="xunit.abstractions" version="2.0.1" targetFramework="net45" />
-  <package id="xunit.assert" version="2.2.0-beta2-build3300" targetFramework="net45" />
-  <package id="xunit.core" version="2.2.0-beta2-build3300" targetFramework="net45" />
-  <package id="xunit.extensibility.core" version="2.2.0-beta2-build3300" targetFramework="net45" />
-  <package id="xunit.extensibility.execution" version="2.2.0-beta2-build3300" targetFramework="net45" />
-  <package id="xunit.runner.visualstudio" version="2.2.0-beta2-build1149" targetFramework="net45" developmentDependency="true" />
-</packages>
