+++ DtronixMessageQueue.Tests/Mq/MqServerTests.cs
@@ -1,8 +1,7 @@
-﻿using System;
-using Xunit;
+﻿using Xunit;
 using Xunit.Abstractions;
 
-namespace DtronixMessageQueue.Tests {
+namespace DtronixMessageQueue.Tests.Mq {
 
 
 	public class MqServerTests : MqTestsBase {
