+++ DtronixMessageQueue.Tests/Mq/SimpleMqSession.cs
@@ -1,0 +1,4 @@
+﻿namespace DtronixMessageQueue.Tests.Mq {
+	public class SimpleMqSession : MqSession<SimpleMqSession> {
+	}
+}
