+++ DtronixMessageQueue.Tests/Rpc/Services/Server/ICalculatorService.cs
@@ -1,6 +1,6 @@
-﻿using System;
+﻿using DtronixMessageQueue.Rpc;
 
-namespace DtronixMessageQueue.Rpc.Tests.Services.Server {
+namespace DtronixMessageQueue.Tests.Rpc.Services.Server {
 	public interface ICalculatorService : IRemoteService <SimpleRpcSession>{
 		int Add(int number_1, int number_2);
 		int Subtract(int number_1, int number_2);
