+++ DtronixMessageQueue.Tests/Rpc/SimpleRpcSession.cs
@@ -1,0 +1,6 @@
+﻿using DtronixMessageQueue.Rpc;
+
+namespace DtronixMessageQueue.Tests.Rpc {
+	public class SimpleRpcSession : RpcSession<SimpleRpcSession> {
+	}
+}
