--- DtronixMessageQueue.Tests/SimpleMqSession.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DtronixMessageQueue.Tests {
-	public class SimpleMqSession : MqSession<SimpleMqSession> {
-	}
-}
