+++ DtronixMessageQueue/MqSession.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using DtronixMessageQueue.Socket;
-using NLog;
 
 namespace DtronixMessageQueue {
 
@@ -13,11 +12,6 @@ namespace DtronixMessageQueue {
 	/// </summary>
 	public abstract class MqSession<TSession> : SocketSession
 		where TSession : MqSession<TSession>, new() {
-
-		/// <summary>
-		/// Logger for this class.
-		/// </summary>
-		private static ILogger logger = LogManager.GetCurrentClassLogger();
 
 		/// <summary>
 		/// True if the socket is currently running.
@@ -221,7 +215,7 @@ namespace DtronixMessageQueue {
 		/// <param name="sender">Originator of call for this event.</param>
 		/// <param name="e">Event args for the message.</param>
 		public virtual void OnIncomingMessage(object sender, IncomingMessageEventArgs<TSession> e) {
-			logger.Debug("Session {0}: Received {1} messages.", Id, e.Messages.Count);
+			//logger.Debug("Session {0}: Received {1} messages.", Id, e.Messages.Count);
 			IncomingMessage?.Invoke(sender, e);
 		}
