--- DtronixMessageQueue/NLog.config
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
-		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		xsi:schemaLocation="http://www.nlog-project.org/schemas/NLog.xsd NLog.xsd"
-		autoReload="true"
-		throwExceptions="false"
-		internalLogLevel="Off" internalLogFile="c:\temp\nlog-internal.log" >
-	<targets>
-	 <target xsi:type="NLogViewer"
-			name="viewer"
-			includeSourceInfo="true"
-			address="udp://127.0.0.1:9999"/>
-	</targets>
-
-	<rules>
-		<logger name="DtronixMessageQueue.*" minlevel="Trace" writeTo="viewer" />
-	</rules>
-</nlog>
