+++ DtronixMessageQueue/Rpc/IRemoteService.cs
