+++ DtronixMessageQueue/Rpc/RpcMessageType.cs
