+++ DtronixMessageQueue/Rpc/RpcRemoteException.cs
