+++ DtronixMessageQueue/Rpc/RpcRemoteExceptionDataContract.cs
