+++ DtronixMessageQueue/Rpc/RpcSession.cs
