+++ DtronixMessageQueue/Rpc/SerializationStore.cs
