+++ DtronixMessageQueue/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NLog" version="4.3.8" targetFramework="net45" />
   <package id="NuGet.CommandLine" version="3.4.3" targetFramework="net45" developmentDependency="true" />
+  <package id="protobuf-net" version="2.1.0" targetFramework="net45" />
 </packages>
